+++ ownCloud/Client/ClientRootViewController.swift
@@ -224,12 +224,6 @@ class ClientRootViewController: UITabBarController, UINavigationControllerDelega
 
 	override func viewDidAppear(_ animated: Bool) {
 		super.viewDidAppear(animated)
-		if MediaUploadQueue.shallShowUploadUnfinishedWarning 	(for: self.bookmark) {
-			let unfinishedUploadAlert = UIAlertController(with: "Warning".localized,
-														  message: "Media upload in the previous session was incomplete since the application was terminated".localized)
-			self.present(unfinishedUploadAlert, animated: true, completion: nil)
-			MediaUploadQueue.resetUploadPendingFlag(for: self.bookmark)
-		}
 	}
 
 	var closeClientCompletionHandler : (() -> Void)?
